narrowing fixes: 5 MWERKS-narrowing sentinels made explicit
Fixes surfacing when the parent sunbright build is compiled without
-Wno-narrowing on sms-native. Each fix preserves the exact wire value
MWERKS produced when it silently accepted the narrowing.

- src/Enemy/emario.cpp:216 — case 0x80000001 in a switch(s32) needs
  an explicit s32 cast (2147483649u → -2147483647 signed).
- src/Enemy/hamukuri.cpp:473-476 — GXColorS10 channel initializers
  0xFFEC/0xFFA6/0xFF74/0xFFD8/0xFFBA/0xFFCE are s16 negative values;
  explicit (s16) casts preserve bit pattern.
- src/JSystem/JAudio/JASystem/JASPlayer_impl.cpp:105,108 — s16 vib/tre
  table sentinels 0xC000 / 0x8001 explicit-cast to (s16).
- src/MSound/MSound.cpp:35 — u16[32] sentinel initializer -1 replaced
  with 0xFFFF (the same wire value, GCC-valid).

Not fixed in this commit (parent's sms-native target retains
-Wno-narrowing to cover the remaining ~30 MWERKS-narrowing sites in
this decomp library; a proper eradication needs per-site type-correcting
across ~500 TUs and is out of scope for a first-party build-hygiene
sweep).

+++ src/Enemy/hamukuri.cpp
@@ -470,10 +470,14 @@ TSpineEnemy* TBossDangoHamuKuriManager::createEnemyInstance()
 }
 
 static const GXColorS10 mFireHamNoseColorStart = { 0x1F4, 0xC8, 0x78, 0xFF };
-static const GXColorS10 mFireHamNoseColorEnd = { 0xFFEC, 0xFFA6, 0xFF74, 0xFF };
+// GXColorS10 fields are s16 (signed 10-bit-range channel value). MWERKS treated
+// large hex constants as s16 by narrowing; GCC diagnoses. Explicit s16 casts
+// preserve the exact wire values (0xFFEC = -20, 0xFFA6 = -90, 0xFF74 = -140,
+// 0xFFD8 = -40, 0xFFBA = -70, 0xFFCE = -50).
+static const GXColorS10 mFireHamNoseColorEnd = { (s16)0xFFEC, (s16)0xFFA6, (s16)0xFF74, 0xFF };
 static const GXColorS10 mFireHamOtherColorStart
-    = { 0x1F4, 0xFFD8, 0xFFBA, 0xFF };
-static const GXColorS10 mFireHamOtherColorEnd = { 0x50, 0x5, 0xFFCE, 0xFF };
+    = { 0x1F4, (s16)0xFFD8, (s16)0xFFBA, 0xFF };
+static const GXColorS10 mFireHamOtherColorEnd = { 0x50, 0x5, (s16)0xFFCE, 0xFF };
 
 static GXColorS10 mFireHamNoseColorDiff = {
 	mFireHamNoseColorEnd.r - mFireHamNoseColorStart.r,
